requirements.txt
faker==33.1.0
duckdb==1.1.3
pyyaml==6.0.2
python-dotenv==1.0.1
pandas==2.2.3